# gdrive-aggregator-be/docs/swagger.yaml
basePath: /
definitions:
...
  model.Account:
    properties:
      avatar_url:
        type: string
      created_at:
        type: string
      display_name:
        type: string
      email:
        type: string
      files:
        description: Relations
        items:
          $ref: '#/definitions/model.FileRecord'
        type: array
      id:
        type: integer
      is_active:
        type: boolean
      storage_limit:
        description: bytes
        type: integer
      storage_usage:
        description: bytes
        type: integer
      updated_at:
        type: string
      usage_in_drive:
        type: integer
      usage_in_trash:
        type: integer
    type: object
...
  model.BaseResponse:
    properties:
      data: {}
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
  model.ErrorResponse:
    properties:
      error:
        example: detail error
        type: string
      message:
        example: Terjadi kesalahan
        type: string
      success:
        example: false
        type: boolean
    type: object
  model.FileRecord:
    properties:
      account_email:
        type: string
      account_id:
        type: integer
      created_at:
        type: string
      drive_file_id:
        type: string
      icon_link:
        type: string
      id:
        type: integer
      is_folder:
        type: boolean
      md5_checksum:
        type: string
      mime_type:
        type: string
      name:
        type: string
      parent_id:
        type: string
      size:
        description: bytes
        type: integer
      updated_at:
        type: string
      web_view_link:
        type: string
    type: object
...
info:
  contact:
    name: Google Drive Aggregator Support
  description: |-
    REST API untuk mengagregasi beberapa akun Google Drive ke dalam satu storage pool terpusat.
    Mendukung smart allocation (otomatis pilih akun dengan sisa kuota terbesar), streaming download, dan manajemen multi-akun.
  title: Google Drive Aggregator API
  version: "1.0"
paths:
  /api/v1/accounts:
    get:
      consumes:
      - application/json
      description: Menampilkan semua akun Google yang terhubung ke aggregator
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.BaseResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/model.Account'
                  type: array
              type: object
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Daftar Semua Akun Terhubung
      tags:
      - Accounts
  /api/v1/accounts/{id}:
    delete:
      consumes:
      - application/json
      description: Menghapus akun Google dari aggregator
      parameters:
      - description: ID Akun
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.BaseResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Hapus / Putuskan Akun
      tags:
      - Accounts
    get:
      consumes:
      - application/json
      description: Menampilkan detail akun Google tertentu berdasarkan ID
      parameters:
      - description: ID Akun
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Account'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Detail Akun
      tags:
      - Accounts
  /api/v1/accounts/{id}/sync:
    post:
      consumes:
      - application/json
      description: Memperbarui data penggunaan storage akun langsung dari Google Drive
        API
      parameters:
      - description: ID Akun
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Account'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Sinkronisasi Kuota Akun
      tags:
      - Accounts
  /api/v1/auth/google/callback:
    get:
      consumes:
      - application/json
      description: Menerima authorization code dari Google dan menyimpan token akun
        ke database
      parameters:
      - description: Authorization code dari Google
        in: query
        name: code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/model.BaseResponse'
            - properties:
                data:
                  $ref: '#/definitions/model.Account'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ErrorResponse'
      summary: Callback Google OAuth2
      tags:
      - Auth
...
swagger: "2.0"